app/requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=8.0.0
ultralytics>=8.0.0
PyQt5>=5.15.0
scipy>=1.7.0 